feat: added author controller
