+++ routes/auth.js
@@ -1,0 +1,9 @@
+const express = require("express");
+const controller = require("../controllers/auth");
+
+const router = express.Router();
+
+router.post("/login", controller.login);
+router.post("/register", controller.register);
+
+module.exports = router;
