+++ routes/author.js
@@ -1,9 +1,10 @@
 const express = require("express");
+const passport = require("passport")
 const controller = require("../controllers/author");
 
 const router = express.Router();
 
-router.post("/login", controller.login);
-router.post("/register", controller.register);
+router.get("/", passport.authenticate('jwt', {session: false}), controller.getAuthor);
+router.patch("/:id", passport.authenticate('jwt', {session: false}), controller.update);
 
 module.exports = router;
